Reorganize crate, fix test
