+++ gtest/src/manager/reservations.rs
@@ -18,18 +18,13 @@
 
 //! Various reservation related methods for ExtManager
 
+use super::ExtManager;
 use gear_common::{
-    auxiliary::BlockNumber,
-    gas_provider::Imbalance,
     scheduler::{ScheduledTask, StorageType},
     storage::Interval,
-    MessageId, Origin, ProgramId, ReservationId,
+    ProgramId, ReservationId,
 };
-use gear_core::{pages::num_traits::Zero, reservation::GasReservationSlot};
-
-use crate::RESERVE_FOR;
-
-use super::ExtManager;
+use gear_core::reservation::GasReservationSlot;
 
 impl ExtManager {
     pub(crate) fn remove_reservation(
@@ -103,69 +98,5 @@ impl ExtManager {
         self.consume_and_retrieve(reservation);
 
         slot
-    }
-
-    pub(crate) fn consume_and_retrieve(&mut self, id: impl Origin) {
-        let outcome = self.gas_tree.consume(id).unwrap_or_else(|e| {
-            let err_msg = format!(
-                "consume_and_retrieve: failed consuming the rest of gas. Got error - {e:?}"
-            );
-
-            unreachable!("{err_msg}")
-        });
-
-        if let Some((imbalance, multiplier, external)) = outcome {
-            let gas_left = imbalance.peek();
-
-            if !gas_left.is_zero() {
-                self.bank
-                    .withdraw_gas(external.cast(), gas_left, multiplier);
-            }
-        }
-    }
-
-    pub(crate) fn charge_for_hold(
-        &mut self,
-        id: impl Origin,
-        hold_interval: Interval<BlockNumber>,
-        storage_type: StorageType,
-    ) {
-        let id: MessageId = id.cast();
-        let current = self.block_height();
-
-        // Deadline of the task.
-        let deadline = hold_interval.finish.saturating_add(RESERVE_FOR);
-
-        // The block number, which was the last paid for hold.
-        //
-        // Outdated tasks can end up being store for free - this case has to be
-        // controlled by a correct selection of the `ReserveFor` constant.
-        let paid_until = current.min(deadline);
-
-        // holding duration
-        let duration: u64 = paid_until.saturating_sub(hold_interval.start).into();
-
-        // Cost per block based on the storage used for holding
-        let cost = Self::cost_by_storage_type(storage_type);
-
-        let amount = storage_type.try_into().map_or_else(
-            |_| duration.saturating_mul(cost),
-            |lock_id| {
-                let prepaid = self.gas_tree.unlock_all(id, lock_id).unwrap_or_else(|e| {
-                    let err_msg = format!(
-                        "charge_for_hold: failed unlocking locked gas.
-                        Got error - {e:?}"
-                    );
-
-                    unreachable!("{err_msg}");
-                });
-
-                prepaid.min(duration.saturating_mul(cost))
-            },
-        );
-
-        if !amount.is_zero() {
-            self.spend_gas(id, amount);
-        }
     }
 }
